feat: Created Vertical Stepper Component
Implemented a vertical stepper component with customizable styles for steps, lines, and overall layout. Added support for step labels, progress indicators, and click handling. Also includes utility functions and model classes for styling and layout configurations.
