+++ kotstep/src/main/java/com/binayshaw7777/kotstep/v2/component/layout/HorizontalKotStep.kt
@@ -1,0 +1,16 @@
+package com.binayshaw7777.kotstep.v2.component.layout
+
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import com.binayshaw7777.kotstep.v2.model.step.Step
+import com.binayshaw7777.kotstep.v2.model.style.KotStepStyle
+
+@Composable
+fun HorizontalKotstep(
+    modifier: Modifier = Modifier,
+    currentStep: () -> Float,
+    style: KotStepStyle,
+    steps: List<Step>
+) {
+
+}
