+++ kotstep/src/main/java/com/binayshaw7777/kotstep/v2/model/style/IconStyle.kt
@@ -1,0 +1,10 @@
+package com.binayshaw7777.kotstep.v2.model.style
+
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.unit.Dp
+import androidx.compose.ui.unit.dp
+
+data class IconStyle(
+    val iconTint: Color = Color.Unspecified,
+    val iconSize: Dp = 16.dp
+)
