+++ kotstep/src/main/java/com/binayshaw7777/kotstep/v2/model/style/LineType.kt
@@ -1,0 +1,7 @@
+package com.binayshaw7777.kotstep.v2.model.style
+
+enum class LineType {
+    SOLID,
+    DOTTED,
+    DASHED
+}
