+++ kotstep/src/main/java/com/binayshaw7777/kotstep/v2/util/Util.kt
@@ -1,6 +1,11 @@
 package com.binayshaw7777.kotstep.v2.util
 
+import androidx.compose.foundation.clickable
+import androidx.compose.foundation.interaction.MutableInteractionSource
 import androidx.compose.foundation.layout.PaddingValues
+import androidx.compose.runtime.remember
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.composed
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.unit.dp
 import com.binayshaw7777.kotstep.v2.model.step.StepLayoutStyle
@@ -16,13 +21,14 @@ object Util {
     fun getKotStepStyle(): KotStepStyle {
         return KotStepStyle(
             stepLayoutStyle = StepLayoutStyle.Vertical,
-            isScrollable = true,
+            isScrollable = false,
             showCheckMarkOnDone = true,
             ignoreCurrentState = false,
             stepStyle = StepStyles.default().copy(
                 onTodo = StepStyle.defaultTodo().copy(
                     stepSize = 50.dp,
                     stepColor = Color.Gray,
+                    borderStyle = BorderStyle(width = 2.dp, color = Color.Red)
                 ),
                 onCurrent = StepStyle.defaultTodo().copy(
                     stepSize = 60.dp,
@@ -32,23 +38,34 @@ object Util {
                 onDone = StepStyle.defaultTodo().copy(
                     stepSize = 50.dp,
                     stepColor = Color.Green,
+                    borderStyle = BorderStyle(width = 2.dp, color = Color.DarkGray)
                 )
             ),
             lineStyle = LineStyles.default().copy(
                 onTodo = LineStyle.defaultTodo().copy(
-                    lineThickness = 4.dp,
-                    lineLength = 20.dp
+                    lineThickness = 10.dp,
+                    lineLength = 100.dp,
+                    linePadding = PaddingValues(2.dp)
                 ),
                 onCurrent = LineStyle.defaultCurrent().copy(
-                    lineThickness = 4.dp,
-                    lineLength = 20.dp,
-                    linePadding = PaddingValues(4.dp)
+                    lineThickness = 10.dp,
+                    lineLength = 100.dp,
+                    linePadding = PaddingValues(2.dp)
                 ),
                 onDone = LineStyle.defaultDone().copy(
-                    lineThickness = 4.dp,
-                    lineLength = 20.dp
+                    lineThickness = 10.dp,
+                    lineLength = 100.dp,
+                    linePadding = PaddingValues(2.dp)
                 )
             )
         )
+    }
+
+    fun Modifier.onClick(onClick: () -> Unit): Modifier = composed {
+        this.clickable(
+            indication = null,
+            interactionSource = remember { MutableInteractionSource() }) {
+            onClick()
+        }
     }
 }
